Use int for message IDs in JobEntity

+++ Database/Models/JobEntity.cs
@@ -1,4 +1,4 @@
-using Devil7Softwares.TeraboxDownloader.Downloader;
+using Devil7Softwares.TeraboxDownloader.Terabox;
 using Devil7Softwares.TeraboxDownloader.Enums;
 
 namespace Devil7Softwares.TeraboxDownloader.Database.Models;
@@ -11,9 +11,9 @@ internal class JobEntity
 
     public virtual ChatEntity? Chat { get; set; }
 
-    public long MessageId { get; set; }
+    public int MessageId { get; set; }
 
-    public long StatusMessageId { get; set; }
+    public int StatusMessageId { get; set; }
 
     public string Url { get; set; } = string.Empty;
